requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
WTForms==3.0.1
email-validator==2.1.1
qrcode==7.4
Pillow==10.0.0
reportlab==4.0.9
python-dotenv==1.0.0
pyotp
itsdangerous
waitress
flask-migrate
